mqtt-consumer/requirements.txt
paho-mqtt==1.6.1
confluent-kafka==1.9.2